requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface 
faiss-cpu
langchain-ollama
numpy==1.26.4
chromadb
langchain_chroma
streamlit